qcommon: fix travis build

+++ src/qcommon/q_shared.h
@@ -647,7 +647,7 @@ void Parse1DMatrix(char **buf_p, int x, float *m);
 void Parse2DMatrix(char **buf_p, int y, int x, float *m);
 void Parse3DMatrix(char **buf_p, int z, int y, int x, float *m);
 
-int QDECL Com_sprintf(char *dest, size_t size, const char *fmt, ...) __attribute__ ((format(printf, 3, 4)));
+int QDECL Com_sprintf(char *dest, unsigned int size, const char *fmt, ...) __attribute__ ((format(printf, 3, 4)));
 
 char *Com_SkipTokens(char *s, int numTokens, const char *sep);
 char *Com_SkipCharset(char *s, char *sep);
